requirements.txt
fastapi
Pillow
python-multipart
transformers
uvicorn
motor
pydantic[email]
passlib
python-jose
python-dotenv